requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
email-validator
python-slugify
alembic
razorpay
stripe
fastapi-csrf-protect
slowapi
starlette
bleach
sqlalchemy-utils
python-dotenv
itsdangerous
pytest
httpx
pydantic-settings